cv/recognize-license-plate/requirements.txt
google-cloud-storage==1.26.0